Add CI job to build and run services
This is the first step towards e2e testing. For now we will start by
building, and running, the various services. In the mean time, we are
also working on getting proper e2e testing to run in CI. A few changes
are still needed for the later to happen but, this PR should get us a
step closer and to a better place.

related to #110

Signed-off-by: Flavio Percoco <flaper87@gmail.com>