Add an stdout output
The original strategy for the `stdout` logger was to simply add
`os.Stdout` to the outputs enabled. This worked well for the
non-containerazed case. When running within the container, os.Stdout was
not properly flushing the buffer, causing the log to never be printed in
stdout.

Instead of playing too much with it and doing some container-specific
hacks, this commits adds a new stdout logger which has the benefit of
being consisted with the rest of the loggers. This commit appends a
new-line character at the end of the byte array passed to `os.Stdout` to
guarantee one log will be printed per line and to force flush of stdout
in the container.

In addition, we can now set custom settings for this logger, if needed.

This commit also moves the `initializer.go` module to the root `pkg` so
that we can move the `Config` struct to `config.go` and use `log_level`
as well as other global configs during the initialization of the
service. This preserves the behavior from before the refactor.

Signed-off-by: Flavio Percoco <flavio@reblaze.com>